agents/strict_auditor_juror/pyproject.toml
[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[project]
name = "strict-auditor-juror-agent"
version = "1.0.0"
description = "Strict Auditor Juror Agent - Prosecutor Perspective"
requires-python = ">=3.10"

dependencies = [
    "google-adk>=1.0.0",
    "google-cloud-aiplatform>=1.40.0",
    "google-genai>=0.3.0",
]
